Update the `qk-denote-find-notes` function, don't need affe
If we just go back and think what we were trying to do with the
affe-find function is just to have a list of files.

Since we are trying to be as lean as possible, using find-file is the
leanest and fastest way we can go about it. In order to have the
current-directory be the denote-directory, we use the new macro with-current-directory!