Add variables to optimize org-agenda generation
After some time, if you don't take the time to archive your DONE items
for the projects, the agenda will take more and more time to build.

I wonder whether the lenght of the files is also interesting, I would
say the general guidelines for this is to make small project files that
we archive every two days (we should have enough for one standup
meeting) and an investigation file for each of the projects.

https://orgmode.org/worg/agenda-optimization.html