Implement archive to dailies workflow with org TODO entries
The org-agenda bulks up with DONE tasks, and I don't really want to
take the time and archive them individually. I would rather archive them
to a location for the specific day (say, a daily), and be able to come
back in time for the tasks that I have been doing for everyday.

This approach would give us the tools to create an org-agenda view with
only the data from the dailies. It would be expensive but it shouldn't
be an issue once it's loaded once, the cache works great.

Also, created a command to access the dailies folder, to check easily
the tasks for a single day or today.