Add qk-denote--migrate to move all the org-roam files to denote
It's using the `qk-denote--rename-file-on-tags-change` capabilities and
visits each of the files. On visit, we have the function changing the
tags if not the same as the filetags+ on the file.

After the migration, I removed the find-file hook for the function,
since it's not polished enough and it breaks denote-org-capture. I'll
try to come back and fix that and add to the find-file hook again.