Refactor creation of org-agenda with denote
Denote creates the org-agenda with an async function, calling the rg
cli to get the list of projects. For the org-agenda use-case, we can
just wait for the resulting function and take the result.

For implementing this use case, I created a helper function that calls
our custom org-agenda function interactively. This helper function is
invoked interactively and we can remove the hooks on org-agenda, making
the process seamless.

Drawback: org-agenda(s) takes lot of build time in general. Even though
I have optimized the files that the agenda builds, and the number is
reduced, it still takes some time. I might need to investigate how to
fix this in the future.