Add orderless lsp-mode configuration
Still on the testing phase, I'll update this if I find this to be too
time consuming for the lsp-mode use case. It really shouldn't be.

I tested this with the rust-analyzer lsp and it seemed to do the trick
pretty well. If I wanted to have the SPC orderless separation with
corfu, I would add more config, but I haven't tested it yet.

Finally, also add bulk action on "x" keybinding on the org-agenda, kinda
following the "x" bindings on other modes, like marking on mu4e headers.