Replace persp-mode for the built-in tab-bar
The behaviour is there, but now the doom modeline isn't working as it
should be, showing *invalid* in place of the clock. I wonder whether
this is happening for this change or the changes in the previous commit.

We can switch to a new tab with SPC p s, and we can switch to the
previously accessed tab with SPC p l. The harpoon name is still working
as previously, so we have the harpoon files per-tab.