Fix Anki integration, improve capture and lazy loading
We now lazy load anki just before we are pushing to the server. With
that, we can avoid any problems that might arise if we have the function
not already required using :general, which was a problem previously.

Also, add the qk-org-capture-here binding, to `SPC m h`, on org-mode
buffers. Keep in mind the binding `SPC o h` that runs consult over the
list of org headings on the current org-mode buffer.