Add qk-refile-save-project-path build-step function
This function will lead to further workflows of getting all the headers
inside a project. Given the current dailies workflow, project headers
are currently empty normally, stored in each of the daily files.

To have the possibility to search all the headers inside a project, the
`qk-refile-save-project-path` sets the header it was refiled to on refile,
which is performant.

Additionally, add the `qk-org-toggle-properties` function, which hides
the properties on the buffers after the refile. I like that, to avoid
the visual clutter of having the 3-5 lines of :PROPERTIES: per header.