Update standup agenda, taking Mondays into account
We now check the day of the week. If we are talking about a monday, or
rather, yesterday was a "Sunday", we build the agenda with Friday and
Monday.

On the process, we put the start day for the agenda to Friday, to avoid
missing information on the initial screen.