Fix `K` binding for context-aware help
If not on eglot, we are going to run the dwim command, that for now just
runs `helpful-dwim`. In the future, we can extend that to other
functionality, which could also be interesting for previews, like denote
links or other links like datetime links, pointing to agenda.

We could create childframes to show that information, could be an
interesting idea of a package.