Add `qk-run-dyncomp` function
Using the `dyncomp` CLI tool, you can run project specific aliases, and
also use global configuration files in case you want to default to a
certain value.

An example of this would be setting a `dyncomp.json` configuration file
on the project root with:

```
{
    "commands": {
        "run": "echo 'hello world'"
    }
}
```

Running `dyncomp run` would now run `echo 'hello world'` wherever on the
project. Added some bindings to that, to run on prog-mode-map.

- "SPC m r": `dyncomp run`.
- "SPC m t": `dyncomp test`.

Can obviously be extended when the need arises. Also created a function
to run dyncomp interactively, with completion with consult, using
compile-history: `qk-run-dyncomp`.