Fix qk-tab-bar and migrate to elpaca-use-package
There was a problem with the initialization of the functions. It seems
like `after!` was breaking the get current tab functions when used in
functions to be auto-loaded.

The fix was to inline getting the current tab name, which we didn't
really need to have abstracted anyway. Tested that it works correctly
now, by creating the tab just after start-up and using the `SPC p l`
binding which calls the `qk-tab-bar-switch-to-last-tab` function, we
were able to get the first tab created, which gets the name from the
buffer name as a default value.