Add qk-constants module, and migrate until qk-completion
Some of the variables have been abstracted into the qk-constants file,
with the intention of reducing the amount of conflict when maintaining
your config starting off this one.

With the constants file and the init.el file for on / off of modules, we
reduce the number of possible conflicted files to two.