Add qk-dired module, to charge dired in a treemacs-like way
We have the possibility to open a dired window for the current
project, with the `SPC p D` key. On that window, you have icons and you
can open the different folders using `tab`.

Additionally, added the `qk-dired-preview-other-window` function to
preview the file while keeping the cursor on th e dired window, for past
project scanning.