Use Vterm as my daily driver terminal emulator
Here's my workflow, that I created on this qk-emacs
distribution. Whenever I want to have the terminal I press MEH-t, which
switches to the other tab named `qk-tab`, which for me is "vterm".

Whenever I want to edit back, I hit MEH-e, and it switches to the
previous tab other than vterm. That is, if we are on vterm, we switch,
if we are already on an "emacs" tab, we stay where we are.