Add new weekly and monthly agenda views with dailies
Add `qk-denote-dailies-between-times` function to standardize the way
that we pull the dailies that are between two dates. You can see the
examples for the monthly and the weekly agenda functions to create your own.

My workflow for standup meetings (daily meetings where we share
progress) is to use the `qk-denote-open-weekly-standup-agenda` function,
bound to `SPC o s` ([O]rg Agenda [S]tandup). The function opens a
org-agenda view with all the DONE items from the past week, which is
handy for reporting.

Addionally, whenever I do my monthly reviews, I would call the
`qk-denote-open-monthly-agenda` (not bound to any key by default, but
you can bind it if you like) and navigate through the weeks with `[[` or
`]]`.

Fixed the issue we had with the `SPC o a` command to open the agenda for
today, that took a long time and bugged Emacs by opening the 700+ files
that I have on my qk-denote-directory. The root cause was that `rg`
returned an empty line (""), which then conflicted with org and opened
all the files for org-agenda creation, which is just too much.