Add bufler and qk-markdown-insert-link-dwim
The later is the same version of our org insert link dwim, but for
markdown-mode, so pressing `SPC m l` will just add the link that we have
on the clipboard, convenient.

Added bufler as well, which is a slightly nicer version of ibuffer. I
really just use it for the project.el integration.