Add org-timeblock and fix prettier
Prettier now is able to format with the prettierc or other configuration
file types in all the cascading directories. This is great for working
with teams that have the configuration files for formatting on the main
directory of the packages.

Org-timeblock is great for knowing at a glance the meetings and other
timeblocks I have in my calendar. Since I have synced my outlook with
org-mode, I see the meetings that I have accepted there. I should have
all meetings though.